Add Docker image for tools volume with modules
This is the docker image for mounted tools volumes and Environment
Modules to load tools into the environment.

Use it with `-v /tools:/tools` to map the local folder /tools to the
tool volume. The tool volume is set up using the librecores
toolsbuilder docker images and must contain the `modulefiles` folder
to work properly.